feat(handler) use the same header type which was received
This change makes the plugin use whatever headers it received: if it received a B3 single header, it will pass on a B3 single header. If it receives a w3c header, it will pass on a w3c header.

Note that the precedence has been changed from the original request for the case where several tracing headers are present simultaneously. Previously w3c would "win". The new precedence is

b3 single > b3 > w3c